+++ R/prop_diff.R
@@ -105,7 +105,6 @@ s_proportion_diff <- function(df,
   checkmate::assert_flag(.in_ref_col, null.ok = TRUE)
   checkmate::assert_list(variables, null.ok = TRUE)
   assert_proportion_value(conf_level)
-  checkmate::assert_character(method)
   checkmate::assert_character(weights_method)
   checkmate::assert_atomic(val)